pyproject.toml
[tool.poetry]
name = "ATRI"
version = "0.3.1"
description = "ATRI is an open source bot."
authors = []
readme = "README.md"
[tool.poetry.dependencies]
python = "^3.10.9"
nb-cli = "^1.0.5"
nonebot2 = {extras = ["fastapi"], version = "^2.0.0rc3"}
msgpack = "1.0.3"
nonebot-adapter-onebot = "^2.2.1"
pillow = "^9.4.0"
wikipedia = "^1.4.0"
nonebot-adapter-cqhttp = "^2.0.0b1"
qrcode = "^7.4.2"
aiohttp = "^3.8.4"
nonebot-plugin-autorepeater = "^0.1.4"
nonebot-plugin-who-at-me = "^0.2.8"
nonebot-plugin-code = "^0.0.5"

[tool.poetry.dev-dependencies]

[tool.nonebot]
plugin_dirs = ["plugins"]
plugins = ["nonebot_plugin_AutoRepeater","nonebot_plugin_who_at_me","nonebot_plugin_code"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
